feat!: restructure code, remove apply and par methods on collections, add a service method
