+++ src/main/scala-2/be/broij/zarrow/package.scala
@@ -14,133 +14,10 @@ limitations under the License.*/
 
 package be.broij
 
-import scala.collection.immutable._
-import scala.language.higherKinds
 import scala.Predef.{identity => id}
-import scala.reflect.ClassTag
 import zio._
 
-/**
- * A `ZArrow` describes a mapping from `I` to a computation that requires an `R`
- * and either succeeds with an `O`, fails with an `E`, or dies. In other words,
- * it describes a mapping from `I` to `ZIO[R, E, O]`.
- */
-final case class ZArrow[-I, -R, +E, +O] private (f: I => ZIO[R, E, O]) extends AnyVal
-
-object ZArrow {
-
-  /**
-   * A `ZArrow` that always maps to a `ZIO` that succeeds with a unit value.
-   */
-  val unit: ZArrow[Any, Any, Nothing, Unit] =
-    ZArrow(_ => ZIO.unit)
-
-  /**
-   * Returns a `ZArrow` that always map to a `ZIO` that succeeds with the
-   * inputs.
-   */
-  def identity[I]: ZArrow[I, Any, Nothing, I] =
-    ZArrow((in: I) => ZIO.succeed(in))
-
-  /**
-   * Lifts `f` in a `ZArrow` that maps any `in: I` to a `ZIO` that requires an
-   * `r: R` and succeeds with an `O` which is the result of executing
-   * `f(in)(r)`. When calling `f`, it will catch any exception and translate it
-   * into a dying `ZIO` effect.
-   */
-  def succeed[I, R: Tag, O](f: I => R => O): ZArrow[I, R, Nothing, O] =
-    ZArrow((in: I) => ZIO.service[R].flatMap(env => ZIO.succeed(f(in)(env))))
-
-  /**
-   * Lifts `f` in a `ZArrow` that maps any `in: I` to a `ZIO` that succeeds with
-   * an `O` which is the result of executing `f(in)`. When calling `f`, it will
-   * catch any exception and translate it into a dying ZIO effect.
-   */
-  def succeed[I, O](f: I => O): ZArrow[I, Any, Nothing, O] =
-    succeed(in => _ => f(in))
-
-  /**
-   * Returns a `ZArrow` that maps any input to a `ZIO` that succeeds with `o`.
-   * When evaluating `o`, it will catch any exception and translate it into a
-   * dying `ZIO` effect.
-   */
-  def succeed[O](o: => O): ZArrow[Any, Any, Nothing, O] =
-    succeed(_ => _ => o)
-
-  /**
-   * Lifts `f` in a `ZArrow` that maps any `in: I` to a `ZIO` that requires an
-   * `r: R & R1` and succeeds with an `O` which is the result of executing
-   * `f(in)(r)`. When calling `f`, it will catch any exception and translate it
-   * into a dying `ZIO` effect.
-   */
-  def fromZIO[I, R, R1: Tag, E, O](f: I => R1 => ZIO[R, E, O]): ZArrow[I, R & R1, E, O] =
-    ZArrow((in: I) => ZIO.service[R1].flatMap(env => ZIO.succeed(f(in)(env)).flatten))
-
-  /**
-   * Lifts `f` in a `ZArrow` that maps any `in: I` to a `ZIO[R, E, O]` which is
-   * the result of executing `f(in)`. When calling `f`, it will catch any
-   * exception and translate it into a dying `ZIO` effect.
-   */
-  def fromZIO[I, R, E, O](f: I => ZIO[R, E, O]): ZArrow[I, R, E, O] =
-    ZArrow((in: I) => ZIO.succeed(f(in)).flatten)
-
-  /**
-   * Lifts `zio` in a `ZArrow` that maps any input to `zio`. When evaluating
-   * `zio`, it will catch any exception and translate it into a dying `ZIO`
-   * effect.
-   */
-  def fromZIO[R, E, O](zio: => ZIO[R, E, O]): ZArrow[Any, R, E, O] =
-    ZArrow(_ => ZIO.succeed(zio).flatten)
-
-  /**
-   * Lifts `f` in a `ZArrow` that maps any `in: I` to a `ZIO` that requires an
-   * `r: R` and succeeds with an `O` which is the result of executing
-   * `f(in)(r)`. When calling `f`, it will catch any exception and translate it
-   * into a failing `ZIO` effect.
-   */
-  def attempt[I, R: Tag, O](f: I => R => O): ZArrow[I, R, Throwable, O] =
-    ZArrow((in: I) => ZIO.service[R].flatMap(env => ZIO.attempt(f(in)(env))))
-
-  /**
-   * Lifts `f` in a `ZArrow` that maps any `in: I` to a `ZIO` that succeeds with
-   * an `O` which is the result of executing `f(in)`. When calling `f`, it will
-   * catch any exception and translate it into a failing `ZIO` effect.
-   */
-  def attempt[I, O](f: I => O): ZArrow[I, Any, Throwable, O] =
-    attempt(in => _ => f(in))
-
-  /**
-   * Returns a `ZArrow` that maps any input to a `ZIO` that succeeds with `o`.
-   * When evaluating `o , it will catch any exception and translate it into a
-   * failing `ZIO` effect.
-   */
-  def attempt[O](o: => O): ZArrow[Any, Any, Throwable, O] =
-    attempt(_ => _ => o)
-
-  /**
-   * Lifts `f` in a `ZArrow` that maps any `in: I` to a `ZIO` that requires an
-   * `r: R & R1` and succeeds with an `O` which is the result of executing
-   * `f(in)(r)`. When calling `f`, it will catch any exception and translate it
-   * into a failing `ZIO` effect.
-   */
-  def fromZIOAttempt[I, R, R1: Tag, E <: Throwable, O](f: I => R1 => ZIO[R, E, O]): ZArrow[I, R & R1, Throwable, O] =
-    ZArrow((in: I) => ZIO.service[R1].flatMap(env => ZIO.attempt(f(in)(env)).flatten))
-
-  /**
-   * Lifts `f` in a `ZArrow` that maps any `in: I` to a `ZIO[R, E, O]` which is
-   * the result of executing `f(in)`. When calling `f`, it will catch any
-   * exception and translate it into a failing `ZIO` effect.
-   */
-  def fromZIOAttempt[I, R, E <: Throwable, O](f: I => ZIO[R, E, O]): ZArrow[I, R, Throwable, O] =
-    ZArrow((in: I) => ZIO.attempt(f(in)).flatten)
-
-  /**
-   * Lifts `zio` in a `ZArrow` that maps any input to `zio`. When evaluating
-   * `zio`, it will catch any exception and translate it into a failing `ZIO`
-   * effect.
-   */
-  def fromZIOAttempt[R, E <: Throwable, O](zio: => ZIO[R, E, O]): ZArrow[Any, R, Throwable, O] =
-    ZArrow(_ => ZIO.attempt(zio).flatten)
+package object zarrow {
 
   implicit class ZArrowOps[I, R, E, O](zArrow: ZArrow[I, R, E, O]) {
 
@@ -161,73 +38,6 @@ object ZArrow {
      */
     def apply(in: I): ZIO[R, E, O] =
       zArrow.f(in)
-
-    /**
-     * Applies this `ZArrow` to each element of the `Set[I]` and returns the
-     * results in a new `Set[O]`.
-     */
-    def apply(set: Set[I]): ZIO[R, E, Set[O]] =
-      ZIO.foreach(set)(zArrow.f)
-
-    /**
-     * Applies this `ZArrow` to each element of the `Set[I]` in parallel and
-     * returns the results in a new `Set[O]`.
-     */
-    def par(set: Set[I]): ZIO[R, E, Set[O]] =
-      ZIO.foreachPar(set)(zArrow.f)
-
-    /**
-     * Applies this `ZArrow` to each element of the `Collection[I]` and returns
-     * the results in a new `Collection[O]`.
-     */
-    def apply[Collection[+Element] <: Iterable[Element]](it: Collection[I])(implicit
-      buildFrom: BuildFrom[Collection[I], O, Collection[O]]
-    ): ZIO[R, E, Collection[O]] =
-      ZIO.foreach(it)(zArrow.f)
-
-    /**
-     * Applies this `ZArrow` to each element of the `Collection[I]` in parallel
-     * and returns the results in a new `Collection[O]`.
-     */
-    def par[Collection[+Element] <: Iterable[Element]](it: Collection[I])(implicit
-      buildFrom: BuildFrom[Collection[I], O, Collection[O]]
-    ): ZIO[R, E, Collection[O]] =
-      ZIO.foreachPar(it)(zArrow.f)
-
-    /**
-     * Applies this `ZArrow` to each element of the `Array[I]` and returns the
-     * results in a new `Array[O]`.
-     */
-    def apply(array: Array[I])(implicit classTag: ClassTag[O]): ZIO[R, E, Array[O]] =
-      ZIO.foreach(array)(zArrow.f)
-
-    /**
-     * Applies this `ZArrow` to each element of the `Array[I]` in parallel and
-     * returns the results in a new `Array[O]`.
-     */
-    def par(array: Array[I])(implicit classTag: ClassTag[O]): ZIO[R, E, Array[O]] =
-      ZIO.foreachPar(array)(zArrow.f)
-
-    /**
-     * Applies this `ZArrow`to each element of the `NonEmptyChunk[I]` and
-     * returns the results in a new `NonEmptyChunk[O]`.
-     */
-    def apply(nonEmptyChunk: NonEmptyChunk[I]): ZIO[R, E, NonEmptyChunk[O]] =
-      ZIO.foreach(nonEmptyChunk)(zArrow.f)
-
-    /**
-     * Applies this `ZArrow` to each element of the `NonEmptyChunk[I]` in
-     * parallel and returns the results in a new `NonEmptyChunk[O]`.
-     */
-    def par(nonEmptyChunk: NonEmptyChunk[I]): ZIO[R, E, NonEmptyChunk[O]] =
-      ZIO.foreachPar(nonEmptyChunk)(zArrow.f)
-
-    /**
-     * Applies this `ZArrow` if the argument is non-empty and returns the
-     * results in a new `Option[O]`
-     */
-    def apply(option: Option[I]): ZIO[R, E, Option[O]] =
-      ZIO.foreach(option)(zArrow.f)
 
     /**
      * Returns a `ZArrow` that maps tuples of shape `(in: I, in1: I1)`. When
@@ -496,30 +306,6 @@ object ZArrow {
       ZArrow.attempt(f).andThen(zArrow)
   }
 
-  implicit class TupledZArrowOps[I1, I2, R, E, O1, O2](zArrow: ZArrow[(I1, I2), R, E, (O1, O2)]) {
-
-    /**
-     * Applies this `ZArrow` to the input `in`, returning the `ZIO` that it is
-     * mapped to.
-     */
-    def apply(in: (I1, I2)): ZIO[R, E, (O1, O2)] =
-      zArrow.f(in)
-
-    /**
-     * Applies this `ZArrow` to each element of the `Map[I1, I2]` and returns
-     * the results in a new `Map[O1, O2]`.
-     */
-    def apply(map: Map[I1, I2]): ZIO[R, E, Map[O1, O2]] =
-      ZIO.foreach(map)((k, v) => zArrow((k, v)))
-
-    /**
-     * Applies this `ZArrow` to each element of the `Map[I1, I2]` in parallel
-     * and returns the results in a new `Map[O1, O2]`.
-     */
-    def par(map: Map[I1, I2]) =
-      ZIO.foreachPar(map)((k, v) => zArrow((k, v)))
-  }
-
   implicit class TupledInputZArrow[I, I1, R, E, O](zArrow: ZArrow[(I, I1), R, E, O]) {
 
     /**
@@ -553,4 +339,5 @@ object ZArrow {
     def filterSecond: ZArrow[I, R, E, O1] =
       zArrow.map(_._2)
   }
+
 }
