+++ src/test/scala-3/be/broij/ZArrowSpec.scala
@@ -12,9 +12,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.*/
 
-package be.broij
+package be.broij.zarrow
 
-import be.broij.ZArrow._
 import zio._
 import zio.test._
 import zio.test.Assertion._
@@ -276,7 +275,7 @@ object ZArrowSpec extends ZIOSpecDefault:
           val layer: ULayer[ZArrow[Int, Any, Nothing, Int]] = zArrow.layer
           check(Gen.int) { int =>
             val io = for {
-              zArrow <- ZIO.service[ZArrow[Int, Any, Nothing, Int]]
+              zArrow <- ZArrow.service[Int, Any, Nothing, Int]
               result <- zArrow(int)
             } yield result
             assertZIO(io.provide(layer))(equalTo(int * 2))
@@ -288,74 +287,6 @@ object ZArrowSpec extends ZIOSpecDefault:
           val zArrow = ZArrow.succeed((i: Int) => i * 2)
           check(Gen.int) { int =>
             assertZIO(zArrow(int))(equalTo(int * 2))
-          }
-        },
-        test("maps the Set to the expected Set") {
-          val zArrow = ZArrow.succeed((i: Int) => i * 2)
-          check(Gen.setOf(Gen.int)) { set =>
-            assertZIO(zArrow(set))(equalTo(set.map(_ * 2)))
-          }
-        },
-        test("maps the Collection to the expected Collection") {
-          val zArrow = ZArrow.succeed((i: Int) => i * 2)
-          check(Gen.listOf(Gen.int)) { list =>
-            assertZIO(zArrow(list))(equalTo(list.map(_ * 2)))
-          }
-        },
-        test("maps the Array to the expected Array") {
-          val zArrow = ZArrow.succeed((i: Int) => i * 2)
-          check(Gen.listOf(Gen.int).map(_.toArray)) { array =>
-            assertZIO(zArrow(array))(equalTo(array.map(_ * 2)))
-          }
-        },
-        test("maps the NonEmptyChunk to the expected NonEmptyChunk") {
-          val zArrow = ZArrow.succeed((i: Int) => i * 2)
-          check(genNonEmptyChunkOf(Gen.int)) { chunk =>
-            assertZIO(zArrow(chunk))(equalTo(chunk.map(_ * 2)))
-          }
-        },
-        test("maps the Option to the expected Option") {
-          val zArrow = ZArrow.succeed((i: Int) => i * 2)
-          check(Gen.option(Gen.int)) { maybe =>
-            assertZIO(zArrow(maybe))(equalTo(maybe.map(_ * 2)))
-          }
-        },
-        test("maps the Map to the expected Map") {
-          val zArrow = ZArrow.succeed((k: Int, v: Int) => (k * 2, v * 4))
-          check(Gen.mapOf(Gen.int, Gen.int)) { intMap =>
-            assertZIO(zArrow(intMap))(equalTo(intMap.map((k, v) => (k * 2, v * 4))))
-          }
-        }
-      ),
-      suite(".par")(
-        test("maps the Set to the expected Set") {
-          val zArrow = ZArrow.succeed((i: Int) => i * 2)
-          check(Gen.setOf(Gen.int)) { set =>
-            assertZIO(zArrow.par(set))(equalTo(set.map(_ * 2)))
-          }
-        },
-        test("maps the Collection to the expected Collection") {
-          val zArrow = ZArrow.succeed((i: Int) => i * 2)
-          check(Gen.listOf(Gen.int)) { list =>
-            assertZIO(zArrow.par(list))(equalTo(list.map(_ * 2)))
-          }
-        },
-        test("maps the Array to the expected Array") {
-          val zArrow = ZArrow.succeed((i: Int) => i * 2)
-          check(Gen.listOf(Gen.int).map(_.toArray)) { array =>
-            assertZIO(zArrow.par(array))(equalTo(array.map(_ * 2)))
-          }
-        },
-        test("maps the NonEmptyChunk to the expected NonEmptyChunk") {
-          val zArrow = ZArrow.succeed((i: Int) => i * 2)
-          check(genNonEmptyChunkOf(Gen.int)) { chunk =>
-            assertZIO(zArrow.par(chunk))(equalTo(chunk.map(_ * 2)))
-          }
-        },
-        test("maps the Map to the expected Map") {
-          val zArrow = ZArrow.succeed((k: Int, v: Int) => (k * 2, v * 4))
-          check(Gen.mapOf(Gen.int, Gen.int)) { intMap =>
-            assertZIO(zArrow.par(intMap))(equalTo(intMap.map((k, v) => (k * 2, v * 4))))
           }
         }
       ),
